Moved registers definition to cpp
With inline definition, there are multiple copies of the same members. The linker merges them when compiling a single shared library, but may not do it for multiple libraries. This happens with gcc, resulting in empty registers in `navground_sim_py` for the `core` classes `Behavior`, `BehaviorModulation` and `Kinematics`, which in turn, make the YAML schema wrong.