Added `apply_inits`
Split the initialization of a world by a scenario in two steps: `Scenario::init_world()` and `Scenario::apply_inits()`. This way, the initializers are called *after* the specialized `init_world`, which can be used to create entities that are then customized in the initializers. Users should normally call `make_world`, which performs these two steps automatically.