Add behavior groups (#24)
Behavior groups compute control commands for a sequence of behaviors. The computation is triggered when the first behavior ask for a new command. They can be used to implement centralized behaviors. No modification to the rest of the API.

Also includes a couple of type error correction in the docs and the stubs.